pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "duckietown-lane-seg"
version = "0.1.0"
description = "Duckietown lane segmentation with vanilla U-Net and configurable ablation variants."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.24",
  "Pillow>=10.0",
  "PyYAML>=6.0",
  "torch>=2.2",
  "torchvision>=0.17",
  "tqdm>=4.66",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

